--- apps/site/src/app/(unauthenticated)/[locale]/(marketing)/components/quote/types/quote.ts
@@ -1,9 +1,0 @@
-// Copyright © Todd Agriscience, Inc. All rights reserved.
-
-/**
- * Quote component props
- * @property {boolean} isDark - Whether the theme is dark
- */
-export interface QuoteProps {
-  isDark?: boolean;
-}
