+++ apps/site/src/i18n/message-files.ts
@@ -15,7 +15,6 @@ export const messageFiles = [
   'research',
   'investors',
   'governance',
-  'landing',
   'contact',
   'disclosures',
   'accessibility',
