--- apps/site/src/messages/landing/es.json
@@ -1,27 +1,0 @@
-{
-  "landing": {
-    "pageHeading": {
-      "title": "Creamos granjas sostenibles",
-      "subtitle": "Nuestra misión es garantizar que la agricultura sostenible beneficie a toda la humanidad.",
-      "button": "Explorar investigación"
-    },
-    "sectionContent": {
-      "section1": {
-        "caption": "Áreas de enfoque",
-        "title": "Somos el laboratorio de investigación de frontera dedicado a construir tecnología de base y genética de cultivos para agricultura sostenible."
-      },
-      "articles": {
-        "title1": "Historias",
-        "title2": "Última investigación"
-      },
-      "explore": {
-        "title": "Explorar todas las investigaciones",
-        "button": "Ver investigaciones"
-      },
-      "contactUs": {
-        "title": "Comenzar con Todd",
-        "button": "Contactar"
-      }
-    }
-  }
-}
